Testing Deploy - 01

+++ migrations/20210122042000-create-user.js
@@ -23,10 +23,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       matchCount: {
-        type: Sequelize.FLOAT
+        type: Sequelize.INTEGER
       },
       winCount: {
-        type: Sequelize.FLOAT
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
